# [webkitdirs] Avoid call to `xcodebuild -showsdks` when computing Xcode options