#### c6027549fd071239887c3b7fed40c73958f06119
<pre>
[webkitdirs] Avoid call to `xcodebuild -showsdks` when computing Xcode options
<a href="https://bugs.webkit.org/show_bug.cgi?id=243257">https://bugs.webkit.org/show_bug.cgi?id=243257</a>

Reviewed by Alexey Proskuryakov.

Since we only use the _name_ of the available SDKs, improve performance
by enumerating the SDK directories instead of calling xcodebuild.

This code is relatively hot compared to most build scripts, as it runs
at least once for every `make` invocation.

* Tools/Scripts/webkitdirs.pm:
(unversionedSDKNameFromSDK):
(availableXcodeSDKs):

Canonical link: <a href="https://commits.webkit.org/253224@main">https://commits.webkit.org/253224@main</a>
</pre>