@emw-apple 2022-07-27
Using the command that Make runs to compute arguments, this is a 3.7x speedup on my system:

Before:
```
Benchmark 1: perl -ITools/Scripts -Mwebkitdirs -e 'print XcodeOptionString()'
  Time (mean ± σ):     565.0 ms ±   4.1 ms    [User: 319.1 ms, System: 219.6 ms]
  Range (min … max):   557.8 ms … 571.5 ms    10 runs
```
 
After:
```
Benchmark 1: perl -ITools/Scripts -Mwebkitdirs -e 'print XcodeOptionString()'
  Time (mean ± σ):     162.0 ms ±   2.3 ms    [User: 55.8 ms, System: 68.8 ms]
  Range (min … max):   158.0 ms … 165.5 ms    10 runs
```